fix: move warning before continue confirmation prompt
